Implement velocity and acceleration reporting
If OpenHMD has the API for velocity and acceleration reports,
use them.

This makes it possible to grab objects remotely in Half-Life Alyx,
but still doesn't quite feel right - we might need to report
angular velocity too, or else there is something wrong with the way
velocity and acceleration are being estimated.